+++ src/components/helper/notificationsRequester.js
@@ -1,8 +1,8 @@
 /* eslint-disable implicit-arrow-linebreak */
 import { ipcRenderer } from 'electron';
 import store from '../../renderer/store';
-import * as categories from '../../constants/notificationCategories';
-import {cpuNotification, memoryNotification} from '../../main/slack/slack.js'
+import * as categories from '../../constants/notificationCategories'; 
+// import {cpuNotification, memoryNotification} from '../../main/slack/slack.js'
 // object that holds what notifications have been sent
 const sentNotifications = {};
 let state;
@@ -160,8 +160,8 @@ const checkForNotifications = (
           );
 
             //sends a slack message if CPU or memory passes threshold
-            cpuNotification();
-            memoryNotification();
+            // cpuNotification();
+            // memoryNotification();
 
 
           // calculate time between now and last notification sent time
